feat(routes/storage): upload-side fire-and-forget bisync trigger
REQ-STOR-015 AC4: after a successful R2 PUT via the storage panel,
fan out a bisync trigger to all the user's running sessions so the
new file appears in the container without waiting up to 15 minutes
for the next cadence cycle.

Refactor: extracts the fan-out logic from the inline implementation
in lifecycle.ts into a shared helper at src/lib/sync-fanout.ts. Both
POST /api/sessions/sync (user-driven, foreground) and the upload-side
auto-trigger (background via executionCtx.waitUntil) now call
fanOutBisyncTrigger. Behaviour and per-session result shape unchanged.

Upload trigger wires into:
- POST / (simple upload, after R2 PutObject success)
- POST /complete (multipart upload finalize, after R2
  CompleteMultipartUpload success)

Both use c.executionCtx.waitUntil() so the upload response is not
blocked on the fan-out. The user perceives the upload as instant;
the "Syncing to container" toast (Phase D D4) tracks fan-out status
via /api/container/status polling per session.

No SYNC_MODE pre-filtering at the Worker layer: each container's
entrypoint.sh applies its own bisync filters, and an "extra" bisync
on a session where the path is excluded is cheap (~one no-op cycle).
Pre-filtering would require reading per-session SYNC_MODE from DO
storage, which is not worth the complexity.

Hibernation safety: fanOutBisyncTrigger is stateless; per-session
errors are isolated. See sync-fanout.ts header comment for the full
hibernation analysis and AD56's fan-out correctness math.
